% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_variance_matrix.R
\name{adjust_variance_matrix}
\alias{adjust_variance_matrix}
\title{Adjust Variance Matrix Estimates}
\usage{
adjust_variance_matrix(mat, tol = 1e-06)
}
\arguments{
\item{mat}{A variance \code{matrix}.}

\item{tol}{A \code{numeric}, positive lower bound of acceptable eigenvalues.
By default, \code{tol = 1e-6}.}
}
\value{
A \code{list} containing the following elements:
\itemize{
\item \code{mat}: adjusted \code{matrix}
\item \code{ok}: a boolean, if \code{FALSE} matrix is adjusted
\item \code{mat_inv}: inverse of \code{mat}
}
}
\description{
This function adjusts variance matrix estimate if it contains negative
eigenvalues.
}
\keyword{internal}
