% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.Rprobit_cl.R
\name{AIC.Rprobit_cl}
\alias{AIC.Rprobit_cl}
\title{Calculate the AIC Value Based on a Fitted Model}
\usage{
\method{AIC}{Rprobit_cl}(object, ...)
}
\arguments{
\item{object}{An \code{\link{Rprobit_cl}} object.}

\item{...}{Not used.}
}
\value{
A \code{numeric}, the AIC value.
}
\description{
This function calculates the AIC value based on a fitted model.
}
