% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_reference_category.R
\name{best_reference_category}
\alias{best_reference_category}
\title{Find Best Reference Categories for Categorical Variables}
\usage{
best_reference_category(
  data_raw,
  categorical_variables,
  choice_variable,
  batches = NULL
)
}
\arguments{
\item{data_raw}{A \code{\link{data_raw_cl}} object.}

\item{categorical_variables}{A \code{character} vector, containing the names of the categorical variables.}

\item{choice_variable}{A \code{character}, the name of the choice variable.}

\item{batches}{A \code{list} of boolean vectors to specify subsets of the raw data.
By default, \code{batches = NULL}.}
}
\value{
A \code{matrix} containing information on the number of choices made for each
combination of categories and each choice alternative, ordered by recommendation.
}
\description{
This function finds best reference categories for categorical variables,
optimally a combination of categories that has optimal coverage over all
choice alternatives. To be used before creating dummy variables!
}
\keyword{internal}
