% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.Rprobit_cl.R
\name{BIC.Rprobit_cl}
\alias{BIC.Rprobit_cl}
\title{Calculate the BIC Value Based on a Fitted Model}
\usage{
\method{BIC}{Rprobit_cl}(object, ...)
}
\arguments{
\item{object}{An \code{\link{Rprobit_cl}} object.}

\item{...}{Not used.}
}
\value{
A \code{numeric}, the BIC value.
}
\description{
This function calculates the BIC value based on a fitted model.
}
