% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{biv_gen_cdf}
\alias{biv_gen_cdf}
\title{general bivariate normal cdf calculation}
\usage{
biv_gen_cdf(b, Sigma)
}
\arguments{
\item{b}{vector; point to evaluate CDF}

\item{Sigma}{2x2 matrix; variance matrix.}
}
\value{
double; CDF
}
\description{
The function computes the bivariate Gaussian CDF.
}
\keyword{internal}
