% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{biv_normal_cdf}
\alias{biv_normal_cdf}
\title{biv_normal_cdf}
\usage{
biv_normal_cdf(x0, x1, r12)
}
\arguments{
\item{w0}{double; x-coordinate}

\item{w1}{double; y-coordinate}

\item{rho}{double; correlation}
}
\value{
double; cdf
}
\description{
The function computes the bivariate Gaussian CDF.
}
\keyword{internal}
