% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cml_pairs.R
\name{build_cml_pairs}
\alias{build_cml_pairs}
\title{Build CML Weighting Matrices}
\usage{
build_cml_pairs(Rprobit_obj, cml_pair_type = NULL)
}
\arguments{
\item{Rprobit_obj}{An \code{\link{Rprobit_cl}} object.}

\item{cml_pair_type}{A named \code{list} of parameters defining CML pair weights with
\itemize{
\item range parameter \code{"d"},
\item \code{list} of positions of observations \code{"positions"},
\item \code{"weighting_function"} or weighting strategy \code{"name"},
\item individual_weights as a \code{vector} of individual weights.
By default, \code{cml_pair_type = NULL}.
}}
}
\value{
A \code{list} of matrices defining CML pairs and weights.
}
\description{
This function builds a list of matrices to define CML pairs and weights.
}
\keyword{internal}
