% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_derived_system_from_model}
\alias{build_derived_system_from_model}
\title{build_derived_system_from_model}
\usage{
build_derived_system_from_model(theta, mod, time, coord)
}
\arguments{
\item{theta}{parameter vector}

\item{mod}{A \code{\link{mod_StSp_cl}} object.}

\item{time}{vector of observation times}

\item{coord}{integer, indicating with respect to which coordinate the derivative is taken}
}
\value{
matrix dGammaT of derivatives of covariances
}
\description{
Calculates the model based on parameter vector theta and model structure mod
}
