% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_info.R
\name{build_info}
\alias{build_info}
\title{Build Info Part of the Model}
\usage{
build_info(
  Rprobit_obj,
  data_raw,
  form,
  ids,
  name,
  description,
  control,
  info = NULL
)
}
\arguments{
\item{Rprobit_obj}{An \code{\link{Rprobit_cl}} object.}

\item{data_raw}{\code{data_raw}-object}

\item{form}{An object of class \code{formula} of the form
\code{choice ~ Type 1 | Type 2 | Type 3}, where:
\itemize{
\item \code{choice} indicates the choice variable
\item \code{Type 1} regressors are varying over alternatives and enter the model as is
\item \code{Type 2} regressors are constant over alternatives and enter the model with alternative varying coefficients
(setting the coefficient for norming alternative to zero)
\item \code{Type 3} regressors are alternative varying and enter the utilities with alternatives varying coefficients
}}

\item{ids}{string, vector of variable names of the data object, identifying the variables that identify a decision maker}

\item{name}{string, Name for the model}

\item{description}{string, Description for the model}

\item{control}{List, contains control variables for estimation procedure}

\item{info}{List, contains information on the model and estimation performance}
}
\value{
An \code{\link{Rprobit_cl}} object.
}
\description{
This function builds the info part of the model.
}
\keyword{internal}
