% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_par_from_mod.R
\name{build_par_from_mod}
\alias{build_par_from_mod}
\title{Build model parameters from mod-object and compute standard errors}
\usage{
build_par_from_mod(theta, mod, variances = NA, labels = NULL)
}
\arguments{
\item{theta}{A \code{numeric} vector containing model parameters.}

\item{mod}{An object of class \code{\link{mod_cl}}.}

\item{variances}{A \code{matrix} of parameter variances.
By default, \code{variances = NA}, in which case no standard errors are
calculated.}

\item{labels}{A \code{list} with two elements:
\itemize{
\item \code{vars}, a \code{list} of \code{characters}, which are names for the
regressor variables,
\item \code{cats}, a \code{list} of \code{characters}, which are names for the
categories.
}

By default, \code{labels = NULL}, in which case the regressors and categories
are not named.}
}
\value{
A \code{list} containing the following elements:
\item{b}{b}
\item{b_sd}{b_sd}
\item{Omega_chol}{Omega_chol}
\item{Omega}{Omega}
\item{Omega_sd}{Omega_sd}
\item{Sigma_chol}{Sigma_chol}
\item{Sigma}{Sigma}
\item{Sigma_sd}{Sigma_sd}
}
\description{
This function builds model parameters from a \code{mod} object and computes
standard errors using the Delta method.
}
\details{
TODO: The Delta method...
}
\keyword{internal}
