% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_system_from_model_AR_R}
\alias{build_system_from_model_AR_R}
\title{build_system_from_model_R}
\usage{
build_system_from_model_AR_R(theta, mod, time)
}
\arguments{
\item{theta}{parameter vector}

\item{mod}{A \code{\link{mod_StSp_cl}} object.}

\item{time}{vector of observation times}
}
\value{
a structure containing beta, Omega, Sigma.
}
\description{
put evaluations back to R
}
