% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_unbalanced_panel_weights.R
\name{build_unbalanced_panel_weights}
\alias{build_unbalanced_panel_weights}
\title{Calculate unbalanced-panel weights}
\usage{
build_unbalanced_panel_weights(Rprobit_obj)
}
\arguments{
\item{Rprobit_obj}{\code{Rprobit_obj}-object}
}
\value{
dataframe with weights depending on number of observations
}
\description{
Function that calculates unbalanced-panel weights from Rprobit_obj object
}
\keyword{internal}
