% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_emp_like.r
\name{calc_emp_like}
\alias{calc_emp_like}
\title{Find the empirical likelihood value at CMLE for model selection}
\usage{
calc_emp_like(Rprobit_obj)
}
\arguments{
\item{Rprobit_obj}{An \code{\link{Rprobit_cl}} object.}
}
\value{
A real number, the maximum of the empirical likelihood criterion function.
}
\description{
This function estimates the Lagrangian parameter in the empirical likelihood framework and returns the corresponding EL value.
}
