% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_delta.R
\name{calculate_delta}
\alias{calculate_delta}
\title{Build a covariance matrix and calculates the standard deviations}
\usage{
calculate_delta(HL, fL, thL, alt, vv = NULL)
}
\arguments{
\item{HL}{Description}

\item{fL}{Description}

\item{thL}{Description}

\item{alt}{Number of alternatives}

\item{vv}{Covariance of estimates}
}
\value{
A list containing the following elements:
\item{Sigma}{Built matrix}
\item{Sigma_sd}{Standard deviations}
}
\description{
Function that builds a covariance matrix and calculates the standard deviations using the Delta method.
}
\keyword{internal}
