% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_deriv_observability}
\alias{calculate_deriv_observability}
\title{derivative of the observability matrix.}
\usage{
calculate_deriv_observability(A, C, dA, dC, dtime)
}
\arguments{
\item{A}{Eigen::MatrixXd; matrix describing the dynamics.}

\item{C}{Eigen::MatrixXd; matrix describing the relation between state and output.}

\item{dA}{Eigen::MatrixXd; derivative of the matrix describing the dynamics.}

\item{dC}{Eigen::MatrixXd; derivative of the matrix describing the relation between state and output.}

\item{dtime}{Eigen::VectorXd; vector of integers, contains the time lags between observations.}
}
\value{
Eigen::MatrixXd; derivative of the observability matrix.
}
\description{
Computes the derivative of the observability matrix.
}
\keyword{internal}
