% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_grid_probs.R
\name{calculate_grid_probs}
\alias{calculate_grid_probs}
\title{Calculate the point masses at the grid points}
\usage{
calculate_grid_probs(mod, theta)
}
\arguments{
\item{mod}{A \code{mod_cl} object}

\item{theta}{A parameter vector (size depends on mod_cl).}
}
\value{
A real vector of probabilities
}
\description{
This function is a helper function used for the evaluation of the non-parametric mixing distributions.
}
\keyword{internal}
