% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_observability}
\alias{calculate_observability}
\title{observability matrix.}
\usage{
calculate_observability(A, C, dtime)
}
\arguments{
\item{A}{Eigen::MatrixXd; matrix describing the dynamics.}

\item{C}{Eigen::MatrixXd; matrix describing the relation between state and output.}

\item{dtime}{Eigen::VectorXd; vector of integers, contains the time lags.}
}
\value{
Eigen::MatrixXd; observability matrix.
}
\description{
Computes the observability matrix.
}
\keyword{internal}
