% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_allvars.R
\name{check_allvars}
\alias{check_allvars}
\title{Check alignment between 'allvars' and 'data_raw'}
\usage{
check_allvars(data_raw, allvars, alt)
}
\arguments{
\item{data_raw}{\code{data_raw}-object}

\item{allvars}{\code{allvars}-object}

\item{alt}{Number of choice alternatives}
}
\value{
Number of choice alternatives
}
\description{
Function that checks the alignment between 'allvars' and 'data_raw' and returns the found number of alternatives.
}
\keyword{internal}
