% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_avail.R
\name{check_avail}
\alias{check_avail}
\title{Check availability of alternatives}
\usage{
check_avail(data_raw, avail, Rprobit_obj)
}
\arguments{
\item{data_raw}{\code{data_raw}-object}

\item{avail}{List, contains information on the availability for each alternative and each choice occasion}

\item{Rprobit_obj}{\code{Rprobit_obj}-object}
}
\value{
Rprobit_obj
}
\description{
Function that checks the availability of alternatives for each decision maker and choice occasion.
}
\keyword{internal}
