% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cov_mat.R
\name{check_cov_mat}
\alias{check_cov_mat}
\title{Checks if theta vector creates valid covariance matrices}
\usage{
check_cov_mat(theta, mod)
}
\arguments{
\item{theta}{A \code{numeric} vector.}

\item{mod}{A \code{\link{mod_cl}} object.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
This function checks if a valid covariance matrix is created.
}
\keyword{internal}
