% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_raw.R
\name{check_data_raw}
\alias{check_data_raw}
\title{Check if 'data_raw' is a valid 'data_raw'-object}
\usage{
check_data_raw(data_raw, form, ids, allvars)
}
\arguments{
\item{data_raw}{\code{data_raw}-object}

\item{form}{\code{form}-object}

\item{ids}{vector specifying the columns in \code{data_raw} that identify a choice occasion uniquely}

\item{allvars}{\code{allvars}-object}
}
\value{
Number of  alternatives found in \code{data_raw}
}
\description{
Function that checks if 'data_raw' is a valid 'data_raw'-object.
}
\keyword{internal}
