% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mod.R
\name{check_mod}
\alias{check_mod}
\title{Checks if input is a valid 'mod'-object}
\usage{
check_mod(mod = NULL)
}
\arguments{
\item{mod}{\code{mod}-object}
}
\value{
(valid) \code{mod}-object
}
\description{
Function that checks if 'mod' is a valid \code{mod}-object.
}
\keyword{internal}
