% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mod_par.r
\name{check_mod_par}
\alias{check_mod_par}
\title{Checks if parameterisation in the 'mod'-object is ok.}
\usage{
check_mod_par(mod = NULL)
}
\arguments{
\item{mod}{\code{mod}-object}
}
\value{
ok boolean indicating, if \code{mod}-object provides correct parameterisation
}
\description{
Function that checks if matrices in \code{mod}-object have correct dimensions.
}
\keyword{internal}
