% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{choice_probs_nonpara}
\alias{choice_probs_nonpara}
\title{Calculation of choice probabilities for non-parametric models.}
\usage{
choice_probs_nonpara(data_obj, mod, control, cml_pair_type)
}
\arguments{
\item{data_obj}{data_cl object}

\item{mod}{A mod_nonpara_cl object.}

\item{control}{Controls for the probit estimation.}

\item{cml_pair_type}{integer controls the probabilities calculated: -1: single choices, 0: all pairs, 1: adjacent pairs.}
}
\value{
A matrix, containing the choice probabilities for each choice in the rows; for each grid point in the columns.
}
\description{
This function computes the choice probabilities for non-parametric models. Each component is not mixed and hence choice probabilities
relate to each single choice, no parameters are needed, as the grid points are specified in the model object.
}
