% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CML_weights.R
\name{CML_weights}
\alias{CML_weights}
\title{Calculate and Apply CML Weights}
\usage{
CML_weights(Rprobit_obj, control_weights = NULL)
}
\arguments{
\item{Rprobit_obj}{A \code{\link{Rprobit_cl}} object.}

\item{control_weights}{A named \code{list} of control parameters to specify the CML weightings.}
}
\value{
A \code{\link{Rprobit_cl}} object.
}
\description{
This function calculates and applies CML weights for probit modeling.
}
