% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.Rprobit_cl.R
\name{coef.Rprobit_cl}
\alias{coef.Rprobit_cl}
\title{Coefficients of a Fitted Probit Model}
\usage{
\method{coef}{Rprobit_cl}(object, ...)
}
\arguments{
\item{object}{An \code{\link{Rprobit_cl}} object.}

\item{...}{Not used.}
}
\value{
A \code{numeric} vector of coefficients.
}
\description{
This function returns the regression coefficients for a fitted probit model.
}
