% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{commmat}
\alias{commmat}
\title{Returns commutation matrix such that C vec(X) = vec(X')}
\usage{
commmat(r, c)
}
\arguments{
\item{r}{integer; number of rows in the matrix.}

\item{c}{integer; number of columns in matrix.}
}
\value{
matrix
}
\description{
Returns commutation matrix such that C vec(X) = vec(X')
}
\keyword{internal}
