% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_sim_fit.R
\name{compare_sim_fit}
\alias{compare_sim_fit}
\title{Compare true with fitted model}
\usage{
compare_sim_fit(theta_0, Rprobit_obj, show_theta = TRUE)
}
\arguments{
\item{theta_0}{True model parameters for simulated data}

\item{Rprobit_obj}{\code{Rprobit_obj}-object}

\item{show_theta}{boolean, if TRUE (default) compare theta vector, otherwise compare entries of b, Omega and Sigma}
}
\value{
matrix with theta vectors in rows
}
\description{
Function that compares true model parameters of the dgp for a simulated data set with parameters of a fitted model.
}
\keyword{internal}
