% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confmat_Rprobit.R
\name{confmat_Rprobit}
\alias{confmat_Rprobit}
\title{Calculate Confusion Matrix of Model Predictions}
\usage{
confmat_Rprobit(Rprobit_obj, data_new = NULL)
}
\arguments{
\item{Rprobit_obj}{An \code{Rprobit_cl} object}

\item{data_new}{Optionally new data for prediction. Can be
\itemize{
\item a \code{\link{data_cl}} object
\item or a \code{\link{data_raw_cl}} object.
By default, \code{data_new = NULL}.
}}
}
\value{
A \code{list} with elements:
\itemize{
\item \code{pred_choices}, a \code{numeric} vector of predicted choices,
\item \code{choices}, a \code{numeric} vector of actual choices,
\item \code{TT}, a \code{table}, the confusion matrix.
}
}
\description{
This function computes the confusion matrix comparing the predicted choices
with the actual choices.
}
