% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_cl.R
\name{control_cl}
\alias{control_cl}
\title{R6 Object Representing the Estimation Specification}
\description{
A \code{control_cl} object contains the estimation specification.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{control_nlm}}{A \code{list} of arguments controlling \code{\link[stats]{nlm}}.}

\item{\code{probit}}{\itemize{
\item If \code{TRUE}, the full probit likelihood is used for estimation.
\item If \code{FALSE} the CML is used.
}}

\item{\code{approx_method}}{A \code{character}, indicating which method to use for the approximation
of the normal CDF. Can be:
\itemize{
\item \code{"SJ"} (for Solow–Joe approximation)
\item \code{"ME"} (for Mendell–Elston approximation)
\item \code{"TVBS"} (for Two-Variate Bivariate Screening approximation)
}}

\item{\code{hess}}{\itemize{
\item If \code{FALSE}, the calculation of the analytic Hessian is avoided.
\item Else, it is calculated.
}}

\item{\code{pairs_list}}{A \code{list} for using different weighting schemes in the CML.}

\item{\code{el}}{\itemize{
\item If \code{TRUE}, the empirical likelihood is used
\item Else, it is not.
}}

\item{\code{control_weights}}{A \code{list} of weights for the CML.}

\item{\code{control_simulation}}{A \code{list} of instructions for data simulation.}

\item{\code{nCores}}{An \code{integer}, the number of cores to use for parallel computation.}

\item{\code{normalize}}{\itemize{
\item If \code{TRUE}, a normalization of the regressors is added.
\item If \code{FALSE}, it is not.
}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-control_cl-new}{\code{control_cl$new()}}
\item \href{#method-control_cl-set_control_nlm}{\code{control_cl$set_control_nlm()}}
\item \href{#method-control_cl-set_probit}{\code{control_cl$set_probit()}}
\item \href{#method-control_cl-set_normalize}{\code{control_cl$set_normalize()}}
\item \href{#method-control_cl-set_approx_method}{\code{control_cl$set_approx_method()}}
\item \href{#method-control_cl-set_hess}{\code{control_cl$set_hess()}}
\item \href{#method-control_cl-set_el}{\code{control_cl$set_el()}}
\item \href{#method-control_cl-set_nCores}{\code{control_cl$set_nCores()}}
\item \href{#method-control_cl-print}{\code{control_cl$print()}}
\item \href{#method-control_cl-clone}{\code{control_cl$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-control_cl-new"></a>}}
\if{latex}{\out{\hypertarget{method-control_cl-new}{}}}
\subsection{Method \code{new()}}{
initialization function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{control_cl$new(
  control_nlm = list(),
  probit = TRUE,
  approx_method = "SJ",
  hess = FALSE,
  pairs_list = NULL,
  el = FALSE,
  control_weights = NULL,
  control_simulation = NULL,
  nCores = 1,
  normalize = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{control_nlm}}{control list}

\item{\code{probit}}{Boolean whether the full probit likelihood or a CML should be used}

\item{\code{approx_method}}{which approximation method to use. SJ|ME|TVBS.}

\item{\code{hess}}{Boolean; if FALSE calculation of the analytic Hessian is avoided.}

\item{\code{pairs_list}}{used for using different weighting schemes in the CML}

\item{\code{el}}{Boolean whether the empirical likelihood is employed}

\item{\code{control_weights}}{a list of control weights}

\item{\code{control_simulation}}{a list of control options for data simulation}

\item{\code{nCores}}{number of cores to use for parallel calculation.}

\item{\code{normalize}}{Boolean add a normalization of the regressors?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-control_cl-set_control_nlm"></a>}}
\if{latex}{\out{\hypertarget{method-control_cl-set_control_nlm}{}}}
\subsection{Method \code{set_control_nlm()}}{
Set control_nlm
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{control_cl$set_control_nlm(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New control_nlm}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-control_cl-set_probit"></a>}}
\if{latex}{\out{\hypertarget{method-control_cl-set_probit}{}}}
\subsection{Method \code{set_probit()}}{
Set probit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{control_cl$set_probit(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New probit}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-control_cl-set_normalize"></a>}}
\if{latex}{\out{\hypertarget{method-control_cl-set_normalize}{}}}
\subsection{Method \code{set_normalize()}}{
Set normalize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{control_cl$set_normalize(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New normalize}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-control_cl-set_approx_method"></a>}}
\if{latex}{\out{\hypertarget{method-control_cl-set_approx_method}{}}}
\subsection{Method \code{set_approx_method()}}{
Set approx_method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{control_cl$set_approx_method(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New  approx_method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-control_cl-set_hess"></a>}}
\if{latex}{\out{\hypertarget{method-control_cl-set_hess}{}}}
\subsection{Method \code{set_hess()}}{
Set hess
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{control_cl$set_hess(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New hess}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-control_cl-set_el"></a>}}
\if{latex}{\out{\hypertarget{method-control_cl-set_el}{}}}
\subsection{Method \code{set_el()}}{
Set el
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{control_cl$set_el(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New el}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-control_cl-set_nCores"></a>}}
\if{latex}{\out{\hypertarget{method-control_cl-set_nCores}{}}}
\subsection{Method \code{set_nCores()}}{
Set nCores
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{control_cl$set_nCores(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New nCores}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-control_cl-print"></a>}}
\if{latex}{\out{\hypertarget{method-control_cl-print}{}}}
\subsection{Method \code{print()}}{
print object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{control_cl$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-control_cl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-control_cl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{control_cl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
