% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_long_data_raw.R
\name{convert_data_long_data_raw}
\alias{convert_data_long_data_raw}
\title{Convert Long Data}
\usage{
convert_data_long_data_raw(
  df = NULL,
  id_dec = "id_dec",
  id_choice = "id_choice",
  alternative = "mode",
  choice = "choice"
)
}
\arguments{
\item{df}{A \code{data.frame} in long format.
By default, \code{df = NULL}.}

\item{id_dec}{A \code{character}, the column name in \code{df} containing the ids uniquely
identifying each decider.
By default, \code{id_dec = "id_dec"}.}

\item{id_choice}{A \code{character}, the column name in \code{df} identifying the choice situations.
By default, \code{id_choice = "id_choice"}.}

\item{alternative}{A \code{character}, the column name in \code{df} identifying the various
alternatives in the long format.
By default, \code{alternative = "mode"}.}

\item{choice}{A \code{character}, the column name in \code{df} column containing the choices.
By default, \code{choice = "choice"}.}
}
\value{
A \code{\link{data_raw_cl}} object.
}
\description{
This function converts data in long format to an object of class
\code{data_raw_cl}.
}
