% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_raw_cl.R
\name{data_raw_cl}
\alias{data_raw_cl}
\title{R6 Object Representing a Data Object}
\description{
An \code{data_raw_cl} object contains the data for modeling.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{df}}{data.frame containing the raw data}

\item{\code{alt_names}}{strings containing the names of all alternatives.}

\item{\code{id}}{string of the name of the column containing the decider ids.}

\item{\code{choice}}{string of the name of the column containing the choices}

\item{\code{ordered}}{Boolean indicating if choices are ordered.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{N}}{number of deciders}

\item{\code{Tp}}{number of choice decisions per decider}

\item{\code{dec_char}}{decider specific regressors}

\item{\code{varying}}{alternative varying variables}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-data_raw_cl-new}{\code{data_raw_cl$new()}}
\item \href{#method-data_raw_cl-set_df}{\code{data_raw_cl$set_df()}}
\item \href{#method-data_raw_cl-set_alt_names}{\code{data_raw_cl$set_alt_names()}}
\item \href{#method-data_raw_cl-set_id}{\code{data_raw_cl$set_id()}}
\item \href{#method-data_raw_cl-set_choice}{\code{data_raw_cl$set_choice()}}
\item \href{#method-data_raw_cl-set_ordered}{\code{data_raw_cl$set_ordered()}}
\item \href{#method-data_raw_cl-set_dec_char}{\code{data_raw_cl$set_dec_char()}}
\item \href{#method-data_raw_cl-set_varying}{\code{data_raw_cl$set_varying()}}
\item \href{#method-data_raw_cl-print}{\code{data_raw_cl$print()}}
\item \href{#method-data_raw_cl-head}{\code{data_raw_cl$head()}}
\item \href{#method-data_raw_cl-tail}{\code{data_raw_cl$tail()}}
\item \href{#method-data_raw_cl-plot}{\code{data_raw_cl$plot()}}
\item \href{#method-data_raw_cl-clone}{\code{data_raw_cl$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_cl-new"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_cl-new}{}}}
\subsection{Method \code{new()}}{
Create a new data object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_cl$new(
  df = data.frame(),
  alt_names = "",
  id = "",
  choice = "",
  ordered = FALSE,
  varying = "",
  dec_char = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{data.frame containing the raw data}

\item{\code{alt_names}}{strings containing the names of all alternatives.}

\item{\code{id}}{string of the name of the column containing the decider ids.}

\item{\code{choice}}{string of the name of the column containing the choices}

\item{\code{ordered}}{Boolean indicating if choices are ordered.}

\item{\code{varying}}{strings of alternative varying variables}

\item{\code{dec_char}}{strings of variables of decider characteristics}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_cl-set_df"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_cl-set_df}{}}}
\subsection{Method \code{set_df()}}{
Set df
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_cl$set_df(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_cl-set_alt_names"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_cl-set_alt_names}{}}}
\subsection{Method \code{set_alt_names()}}{
Set alt_names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_cl$set_alt_names(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{alt_names}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_cl-set_id"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_cl-set_id}{}}}
\subsection{Method \code{set_id()}}{
Set id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_cl$set_id(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{id}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_cl-set_choice"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_cl-set_choice}{}}}
\subsection{Method \code{set_choice()}}{
Set choice
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_cl$set_choice(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{choice}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_cl-set_ordered"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_cl-set_ordered}{}}}
\subsection{Method \code{set_ordered()}}{
Set ordered
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_cl$set_ordered(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{ordered}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_cl-set_dec_char"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_cl-set_dec_char}{}}}
\subsection{Method \code{set_dec_char()}}{
Set dec_char
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_cl$set_dec_char(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{dec_char}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_cl-set_varying"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_cl-set_varying}{}}}
\subsection{Method \code{set_varying()}}{
Set varying
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_cl$set_varying(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{ordered}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_cl-print"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_cl-print}{}}}
\subsection{Method \code{print()}}{
print object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_cl$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_cl-head"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_cl-head}{}}}
\subsection{Method \code{head()}}{
present head of data frame object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_cl$head()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_cl-tail"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_cl-tail}{}}}
\subsection{Method \code{tail()}}{
present tail of data frame object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_cl$tail()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_cl-plot"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_cl-plot}{}}}
\subsection{Method \code{plot()}}{
plot the data frame: density plot for each chosen alternative.
plot the data set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_cl$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_cl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_cl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_cl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
