% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_raw_StSp_cl.R
\name{data_raw_StSp_cl}
\alias{data_raw_StSp_cl}
\title{R6 Object Representing Model Specifications}
\description{
A \code{data_raw_StSp_cl} object contains a raw data set in the wide format, extending data_raw_cl.
In addition to the data_raw_cl a field 'time' is introduced that lists the times at which samples are taken.
}
\section{Super class}{
\code{\link[Rprobit:data_raw_cl]{Rprobit::data_raw_cl}} -> \code{data_raw_StSp_cl}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{time_col}}{column of data set containing time points}

\item{\code{quest_col}}{column of data set containing number of question}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{time}}{observation time points}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-data_raw_StSp_cl-set_time_col}{\code{data_raw_StSp_cl$set_time_col()}}
\item \href{#method-data_raw_StSp_cl-set_quest_col}{\code{data_raw_StSp_cl$set_quest_col()}}
\item \href{#method-data_raw_StSp_cl-print}{\code{data_raw_StSp_cl$print()}}
\item \href{#method-data_raw_StSp_cl-clone}{\code{data_raw_StSp_cl$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_raw_cl" data-id="head"><a href='../../Rprobit/html/data_raw_cl.html#method-data_raw_cl-head'><code>Rprobit::data_raw_cl$head()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_raw_cl" data-id="initialize"><a href='../../Rprobit/html/data_raw_cl.html#method-data_raw_cl-initialize'><code>Rprobit::data_raw_cl$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_raw_cl" data-id="plot"><a href='../../Rprobit/html/data_raw_cl.html#method-data_raw_cl-plot'><code>Rprobit::data_raw_cl$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_raw_cl" data-id="set_alt_names"><a href='../../Rprobit/html/data_raw_cl.html#method-data_raw_cl-set_alt_names'><code>Rprobit::data_raw_cl$set_alt_names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_raw_cl" data-id="set_choice"><a href='../../Rprobit/html/data_raw_cl.html#method-data_raw_cl-set_choice'><code>Rprobit::data_raw_cl$set_choice()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_raw_cl" data-id="set_class_member"><a href='../../Rprobit/html/data_raw_cl.html#method-data_raw_cl-set_class_member'><code>Rprobit::data_raw_cl$set_class_member()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_raw_cl" data-id="set_dec_char"><a href='../../Rprobit/html/data_raw_cl.html#method-data_raw_cl-set_dec_char'><code>Rprobit::data_raw_cl$set_dec_char()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_raw_cl" data-id="set_df"><a href='../../Rprobit/html/data_raw_cl.html#method-data_raw_cl-set_df'><code>Rprobit::data_raw_cl$set_df()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_raw_cl" data-id="set_id"><a href='../../Rprobit/html/data_raw_cl.html#method-data_raw_cl-set_id'><code>Rprobit::data_raw_cl$set_id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_raw_cl" data-id="set_ordered"><a href='../../Rprobit/html/data_raw_cl.html#method-data_raw_cl-set_ordered'><code>Rprobit::data_raw_cl$set_ordered()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_raw_cl" data-id="set_varying"><a href='../../Rprobit/html/data_raw_cl.html#method-data_raw_cl-set_varying'><code>Rprobit::data_raw_cl$set_varying()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_raw_cl" data-id="tail"><a href='../../Rprobit/html/data_raw_cl.html#method-data_raw_cl-tail'><code>Rprobit::data_raw_cl$tail()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_StSp_cl-set_time_col"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_StSp_cl-set_time_col}{}}}
\subsection{Method \code{set_time_col()}}{
Set time column.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_StSp_cl$set_time_col(time_col)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_col}}{An \code{string}, denoting the column listing the time points}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_StSp_cl-set_quest_col"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_StSp_cl-set_quest_col}{}}}
\subsection{Method \code{set_quest_col()}}{
Set question column.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_StSp_cl$set_quest_col(quest_col)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{quest_col}}{An \code{string}, denoting the column listing the time points}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_StSp_cl-print"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_StSp_cl-print}{}}}
\subsection{Method \code{print()}}{
prints the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_StSp_cl$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_raw_StSp_cl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-data_raw_StSp_cl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_raw_StSp_cl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
