% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_raw_to_data.R
\name{data_raw_to_data}
\alias{data_raw_to_data}
\title{Transform 'data_raw' to 'data'}
\usage{
data_raw_to_data(data_raw, allvars, choice, re, norm_alt, alt = 0, ASC)
}
\arguments{
\item{data_raw}{\code{data_raw_cl}-object}

\item{allvars}{\code{allvars}-object}

\item{choice}{Character name of choice variable}

\item{re}{Names of variables to be considered as random effects}

\item{norm_alt}{Norm alternative: either an integer or a string}

\item{alt}{Number of choice alternatives}

\item{ASC}{Boolean determining whether ASCs are estimated or not}
}
\value{
\code{data_cl} object
}
\description{
Function that takes 'data_raw' and transforms it into 'data'. Operations include expansion of variables into regressors using suitable differencing and omissions to prevent dummy variable trap. Data_raw is assumed to be in wide format, with alternative specific variables indicated using 'var_cat' notation.
}
\keyword{internal}
