% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_StSp_cl.R
\name{data_StSp_cl}
\alias{data_StSp_cl}
\title{R6 Object Representing data}
\description{
A \code{data_StSp_cl} object contains a data set in our specific format, extending data_cl.
In addition to the data_cl a field 'time' is introduced that lists the times at which samples are taken.
}
\section{Super class}{
\code{\link[Rprobit:data_cl]{Rprobit::data_cl}} -> \code{data_StSp_cl}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{time}}{listing the observation times}

\item{\code{quest}}{integer indicating the maximal number of questions per person.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-data_StSp_cl-set_time}{\code{data_StSp_cl$set_time()}}
\item \href{#method-data_StSp_cl-set_quest}{\code{data_StSp_cl$set_quest()}}
\item \href{#method-data_StSp_cl-print}{\code{data_StSp_cl$print()}}
\item \href{#method-data_StSp_cl-clone}{\code{data_StSp_cl$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_cl" data-id="initialize"><a href='../../Rprobit/html/data_cl.html#method-data_cl-initialize'><code>Rprobit::data_cl$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_cl" data-id="plot"><a href='../../Rprobit/html/data_cl.html#method-data_cl-plot'><code>Rprobit::data_cl$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_cl" data-id="set_class_member"><a href='../../Rprobit/html/data_cl.html#method-data_cl-set_class_member'><code>Rprobit::data_cl$set_class_member()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_cl" data-id="set_data"><a href='../../Rprobit/html/data_cl.html#method-data_cl-set_data'><code>Rprobit::data_cl$set_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_cl" data-id="set_ordered"><a href='../../Rprobit/html/data_cl.html#method-data_cl-set_ordered'><code>Rprobit::data_cl$set_ordered()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="data_cl" data-id="set_vars"><a href='../../Rprobit/html/data_cl.html#method-data_cl-set_vars'><code>Rprobit::data_cl$set_vars()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_StSp_cl-set_time"></a>}}
\if{latex}{\out{\hypertarget{method-data_StSp_cl-set_time}{}}}
\subsection{Method \code{set_time()}}{
Set time vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_StSp_cl$set_time(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{A vector, listing the time points}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_StSp_cl-set_quest"></a>}}
\if{latex}{\out{\hypertarget{method-data_StSp_cl-set_quest}{}}}
\subsection{Method \code{set_quest()}}{
Set question number.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_StSp_cl$set_quest(qu)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{qu}}{integer; maximal number of questions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_StSp_cl-print"></a>}}
\if{latex}{\out{\hypertarget{method-data_StSp_cl-print}{}}}
\subsection{Method \code{print()}}{
prints the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_StSp_cl$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_StSp_cl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-data_StSp_cl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{data_StSp_cl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
