% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tr_parallel.R
\name{data_tr_parallel}
\alias{data_tr_parallel}
\title{Transform data for parallel framework}
\usage{
data_tr_parallel(Rprobit_obj, data_tr)
}
\arguments{
\item{Rprobit_obj}{\code{Rprobit_obj}-object}

\item{data_tr}{transformed \code{data}-object}
}
\value{
A list containing the following elements:
\item{Rprobit_obj}{\code{Rprobit_obj}-object}
\item{data_tr_mat_list}{list of transformed data information}
}
\description{
Function that transforms data for parallel framework.
}
\keyword{internal}
