% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlcond}
\alias{dlcond}
\title{Solow-Joe approximation to multivariate Gaussian CDF: gradient calculation}
\usage{
dlcond(x, r)
}
\arguments{
\item{x}{nx1 vector of point to evaluate the CDF at.}

\item{r}{nxn correlation matrix.}
}
\value{
vector; gradient of log of probability.
}
\description{
The function computes the gradient of the CDF for a multivariate Gaussian distribution according to the method of Solow-Joe.
}
\keyword{internal}
