% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlcond_ME}
\alias{dlcond_ME}
\title{Gradient calculation for Mendell-Elston approximation to multivariate Gaussian CDF.}
\usage{
dlcond_ME(x, r)
}
\arguments{
\item{x}{nx1 vector of point to evaluate the CDF at.}

\item{r}{nxn correlation matrix.}
}
\value{
vector; gradient of log of probability.
}
\description{
The function computes the gradient of the CDF for a multivariate Gaussian distribution approximated according to the method of Mendell-Elston.
}
\keyword{internal}
