% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{duplmat}
\alias{duplmat}
\title{duplication matrix}
\usage{
duplmat(n)
}
\arguments{
\item{n}{integer; dimension of matrix.}
}
\value{
matrix
}
\description{
Computes the duplication matrix converting from vech to vec..
}
\keyword{internal}
