% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{elim_ind}
\alias{elim_ind}
\title{find indices for elimination of unwanted rows and columns}
\usage{
elim_ind(n, ind)
}
\arguments{
\item{n}{integer; number of rows of large matrix}

\item{ind}{vector indices.}
}
\value{
an indicator matrix selecting the smaller matrix elements.
}
\description{
Extracts the indices of the entries in the submatrix LSig (ind,ind).
}
