% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{elimmat}
\alias{elimmat}
\title{Returns elimination matrix such that C vec(X) = vech(X)}
\usage{
elimmat(n)
}
\arguments{
\item{n}{integer; dimension of matrix.}
}
\value{
matrix
}
\description{
Returns elimination matrix such that C vec(X) = vech(X)
}
\keyword{internal}
