% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fdiag}
\alias{fdiag}
\title{indices of diagonal elements in vectorized kxk square matrix.}
\usage{
fdiag(k)
}
\arguments{
\item{k}{integer; dimension of matrix.}
}
\value{
matrix
}
\description{
Returns indices of diagonal elements in vectorized kxk square matrix.
}
\keyword{internal}
