% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ASCs_Rprobit.R
\name{fit_ASCs_Rprobit}
\alias{fit_ASCs_Rprobit}
\title{Fit ASCs for an estimated Probit Model}
\usage{
fit_ASCs_Rprobit(Rprobit_obj)
}
\arguments{
\item{Rprobit_obj}{An \code{\link{Rprobit_cl}} object.}
}
\value{
A \code{\link{Rprobit_cl}} object.
}
\description{
This function adjusts the ASCs for an estimated probit model such that
mean predicted frequencies and observed frequencies coincide.
}
