% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_LC_Rprobit.R
\name{fit_LC_Rprobit}
\alias{fit_LC_Rprobit}
\title{Fit a Probit Model including latent classes.}
\usage{
fit_LC_Rprobit(
  Rprobit_obj,
  init_method = "random",
  control_nlm = NULL,
  cml_pair_type = NULL
)
}
\arguments{
\item{Rprobit_obj}{An \code{\link{Rprobit_cl}} object.}

\item{init_method}{A \code{character}, setting the method for initialization of the likelihood
optimization.
Can be one of
\itemize{
\item \code{"random"} (default)
\item or \code{"subset"}.
}}

\item{control_nlm}{Optionally a named \code{list} of control parameters passed on to
\link[stats]{nlm}.
By default, \code{control_nlm = NULL}.}

\item{cml_pair_type}{An integer.
cml_pair_type = 0 (full pairwise), = 1 (adjacent pairwise looped), = 2 (adjacent pairwise CML)
By default, \code{cml_pair_type = NULL}, such that the information in \code{Rprobit_obj} is used.}
}
\value{
A \code{\link{Rprobit_cl}} object.
}
\description{
This function fits a probit model with a number of latent classes.
DOES NOT USE PARALLELIZATION!!!
}
