% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_observation_indices}
\alias{get_observation_indices}
\title{get_observation_indices}
\usage{
get_observation_indices(time, quest, time_n, quest_n)
}
\arguments{
\item{time}{vector of all observation times}

\item{quest}{integer of most questions asked}

\item{time_n}{vector of times when the n-th individual is observed.}

\item{quest_n}{vector of questions answered by the n-th individual.}
}
\value{
a vector including all the observed questions.
}
\description{
Compares the time and question numbers from an observation with the maximal ones and returns a vector containing the indices of observations contained for the
individual.
}
