% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_gen_cdf}
\alias{grad_gen_cdf}
\title{general bivariate normal gradient cdf calculation}
\usage{
grad_gen_cdf(b, Sigma)
}
\arguments{
\item{b}{vector; point to evaluate CDF}

\item{Sigma}{2x2 matrix; variance matrix.}
}
\value{
vector; gradient of CDF
}
\description{
The function computes the gradient of the bivariate Gaussian CDF.
}
\keyword{internal}
