% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad_gen_tvn_cdf}
\alias{grad_gen_tvn_cdf}
\title{general trivariate normal cdf calculation}
\usage{
grad_gen_tvn_cdf(b, Sigma)
}
\arguments{
\item{b}{vector; point to evaluate CDF}

\item{Sigma}{3x3 matrix; variance matrix.}
}
\value{
9x1; gradient of the CDF
}
\description{
The function computes the gradient of the trivariate Gaussian CDF.
}
\keyword{internal}
