% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Hess_gen_cdf}
\alias{Hess_gen_cdf}
\title{general bivariate normal Hessian of cdf calculation}
\usage{
Hess_gen_cdf(b, Sigma)
}
\arguments{
\item{b}{vector; point to evaluate CDF}

\item{Sigma}{2x2 matrix; variance matrix.}
}
\value{
5x5 matrix; Hessian of CDF
}
\description{
The function computes the Hessian of the bivariate Gaussian CDF.
}
\keyword{internal}
