% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Hess_pdf}
\alias{Hess_pdf}
\title{Hess_pdf}
\usage{
Hess_pdf(w0, w1, rho)
}
\arguments{
\item{w0}{double; x-coordinate}

\item{w1}{double; y-coordinate}

\item{rho}{double; correlation}
}
\value{
matrix; Hessian of pdf
}
\description{
The function computes the Hessian of the bivariate Gaussian CDF.
}
\keyword{internal}
