% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Hessian_gen_tvn_cdf}
\alias{Hessian_gen_tvn_cdf}
\title{general trivariate normal cdf calculation}
\usage{
Hessian_gen_tvn_cdf(b, Sigma)
}
\arguments{
\item{b}{vector; point to evaluate CDF}

\item{Sigma}{3x3 matrix; variance matrix.}
}
\value{
9x9; Hessian of the CDF
}
\description{
The function computes the Hessian of the trivariate Gaussian CDF.
}
\keyword{internal}
