% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_cl.R
\name{hypothesis_cl}
\alias{hypothesis_cl}
\title{R6 Object Representing hypothesis Specifications}
\description{
A \code{hypothesis_cl} object contains the specification of a linear hypothesis for a probit model.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Hb}}{A real matrix storing information for restriction beta = Hb*theta + fb.}

\item{\code{fb}}{A real matrix storing information for restriction beta = Hb*theta + fb.}

\item{\code{HO}}{A real matrix storing information for restriction vOm = HO*theta_O + fO.}

\item{\code{fO}}{A real matrix storing information for restriction vOm = HO*theta_O + fO.}

\item{\code{HL}}{A real matrix storing information for restriction vL = HL*theta_L + fL.}

\item{\code{fL}}{A real matrix storing information for restriction vL = HL*theta_L + fL.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hypothesis_cl-new}{\code{hypothesis_cl$new()}}
\item \href{#method-hypothesis_cl-validate}{\code{hypothesis_cl$validate()}}
\item \href{#method-hypothesis_cl-clone}{\code{hypothesis_cl$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hypothesis_cl-new"></a>}}
\if{latex}{\out{\hypertarget{method-hypothesis_cl-new}{}}}
\subsection{Method \code{new()}}{
initialization function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hypothesis_cl$new(
  Hb = matrix(0, 0, 0),
  fb = matrix(0, 0, 0),
  HO = matrix(0, 0, 0),
  fO = matrix(0, 0, 0),
  HL = matrix(0, 0, 0),
  fL = matrix(0, 0, 0)
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Hb}}{matrix}

\item{\code{fb}}{vector}

\item{\code{HO}}{matrix}

\item{\code{fO}}{vector}

\item{\code{HL}}{matrix}

\item{\code{fL}}{vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hypothesis_cl-validate"></a>}}
\if{latex}{\out{\hypertarget{method-hypothesis_cl-validate}{}}}
\subsection{Method \code{validate()}}{
validates if parameters are conformable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hypothesis_cl$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hypothesis_cl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-hypothesis_cl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hypothesis_cl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
