% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{identity}
\alias{identity}
\title{Returns identity matrix}
\usage{
identity(n)
}
\arguments{
\item{n}{integer; dimension of matrix.}
}
\value{
matrix
}
\description{
Returns an identity matrix of the specified dimension
}
\keyword{internal}
