% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_LC_Rprobit.r
\name{init_LC_Rprobit}
\alias{init_LC_Rprobit}
\title{Initialize fit_LC_Rprobit}
\usage{
init_LC_Rprobit(Rprobit_obj, mod, control)
}
\arguments{
\item{Rprobit_obj}{an \code{\link{Rprobit_cl}} object.}

\item{mod}{An \code{\link{mod_cl}} object.}

\item{control}{An \code{\link{mod_cl}} object.}
}
\value{
a vector of model parameters
}
\description{
Function that initializes the estimation routine for latent classes model.
}
