% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_Rprobit.R
\name{init_Rprobit}
\alias{init_Rprobit}
\title{Initialize fit_Rprobit}
\usage{
init_Rprobit(ll_function, Rprobit_obj, init_method, data_tr = NULL)
}
\arguments{
\item{ll_function}{A \code{function} that computes the log-likelihood value.}

\item{Rprobit_obj}{An \code{\link{Rprobit_cl}} object.}

\item{init_method}{A \code{character} specifying the initialization method, one of:
\itemize{
\item \code{"random"}
\item \code{"subset"}
}}

\item{data_tr}{A transformed data set (utilities are differentiated with respect to the
chosen alternative).}
}
\value{
a vector of model parameters
}
\description{
Function that initializes the estimation routine.
}
\keyword{internal}
