% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int2bin}
\alias{int2bin}
\title{convert integer to binary indices}
\usage{
int2bin(num, noBits)
}
\arguments{
\item{num}{integer; to be converted.}

\item{noBits}{integer; number of bits.}
}
\value{
binary vector.
}
\description{
For an integer returns the coding in a binary vector.
}
\keyword{internal}
