% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{int2cats}
\alias{int2cats}
\title{providing categories for integers for cycling through upper and lower bounds}
\usage{
int2cats(num, cat, noBits)
}
\arguments{
\item{num}{integer index}

\item{cat}{number of alternatives}

\item{noBits}{integer number of bits}
}
\value{
vector of categories
}
\description{
converts integers into vector of category indices
}
\keyword{internal}
