% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{J_chol}
\alias{J_chol}
\title{returns the dvech(LL')/dvech(L)}
\usage{
J_chol(x)
}
\arguments{
\item{x}{real vector of vectorization of lower triangular of Cholesky factor}
}
\value{
Jacobian matrix
}
\description{
Returns the Jacobian of the matrix-vectorization with respect to the vech of the Cholesky factor
}
\keyword{internal}
