% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_APU.r
\name{ll_APU}
\alias{ll_APU}
\title{implements the mixture of marginal and pairwise CML function}
\usage{
ll_APU(theta, data_obj, mod, control)
}
\arguments{
\item{theta}{Parameter vector}

\item{data_obj}{A \code{data_cl} object}

\item{mod}{\code{\link{mod_cl}} object encoding the model.}

\item{control}{\code{control_cl} object controlling the evaluation of the likelihood.}
}
\value{
A number containing the attributes "gradient" and "Hessian".
}
\description{
This function adds the marginal CML to the pairwise CML.
}
