% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ll_macml}
\alias{ll_macml}
\title{Calculation of Approximated Composite Marginal Likelihood}
\usage{
ll_macml(theta, data_obj, mod, control)
}
\arguments{
\item{theta}{parameter vector}

\item{data_obj}{data_cl object}

\item{mod}{A \code{\link{mod_cl}} object.}

\item{control}{Controls for the probit estimation.}
}
\value{
A vector, containing the negative log-likelihood with attributes containing the gradient and
(if specified in the controls) the Hessian.
For el==1 the log CML contribution as well as gradients for each decider is contained in the output.
}
\description{
This function computes the composite marginal likelihood.
}
