% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ll_macml_marginal}
\alias{ll_macml_marginal}
\title{Calculation of Approximated Composite Marginal Likelihood}
\usage{
ll_macml_marginal(theta, data_obj, mod, control)
}
\arguments{
\item{theta}{parameter vector}

\item{data_obj}{data_cl object}

\item{mod}{A \code{\link{mod_cl}} object.}

\item{control}{Controls for the probit estimation.}
}
\value{
A vector, containing the negative log-likelihood with attributes containing the gradient and (if specified in the controls) the Hessian.
}
\description{
This function computes the marginal likelihood.
}
