% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ll_macml_norm_factor}
\alias{ll_macml_norm_factor}
\title{Calculation of the normalizing factor for the approximated composite likelihood calculation.}
\usage{
ll_macml_norm_factor(theta, data_obj, mod, control)
}
\arguments{
\item{theta}{parameter vector}

\item{data_obj}{\link{data_cl} object}

\item{mod}{A \code{\link{mod_cl}} object.}

\item{control}{Controls for the probit estimation.}
}
\value{
A vector, containing the negative log-likelihood norming factor.
}
\description{
Computes the norming factor for the CML.
}
\keyword{internal}
