% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ll_macml_o_StSp}
\alias{ll_macml_o_StSp}
\title{MACML function for ordered probit models}
\usage{
ll_macml_o_StSp(theta, data_obj, mod, control)
}
\arguments{
\item{theta}{parameter vector}

\item{data_obj}{\link{data_cl} organised as a data_cl object containing (a) list with elements contains the nxk matrix X and the nx1 vector y (b) N number of deciders (c) number of choice situations per decider}

\item{mod}{A \code{\link{mod_cl}} object.}

\item{control}{Controls for the ordered probit estimation.}
}
\value{
A vector, containing the negative log-likelihood with attributes containing the gradient and (if specified in the controls) the Hessian.
}
\description{
Computes the approximate composite likelihood for the ordered probit case.
}
