% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ll_probit_person}
\alias{ll_probit_person}
\title{Calculation of contribution of one person to the approximated probit log likelihood calculation.}
\usage{
ll_probit_person(X_n, theta, mod, pairwise, approx_method)
}
\arguments{
\item{X_n}{X list of regressor matrices}

\item{theta}{parameter vector}

\item{mod}{A \code{\link{mod_cl}} object.}

\item{pairwise}{integer specifying the pairwise option}

\item{approx_method}{string specifying the approx_method}
}
\value{
A vector, containing the negative log-likelihood with attributes containing the gradient and (if specified in the controls) the Hessian.
}
\description{
Computes the contribution of one person to the log likelihood.
}
\keyword{internal}
