% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.Rprobit_cl.R
\name{logLik.Rprobit_cl}
\alias{logLik.Rprobit_cl}
\title{Provide the Log-Likelihood Value for a Fitted Model}
\usage{
\method{logLik}{Rprobit_cl}(object, ...)
}
\arguments{
\item{object}{An \code{\link{Rprobit_cl}} object.}

\item{...}{Not used.}
}
\value{
A \code{numeric}, the log-likelihood value.
}
\description{
This function provides the log-likelihood value for a fitted model.
}
