% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_id.R
\name{make_id}
\alias{make_id}
\title{Create identifier for decision maker}
\usage{
make_id(data_raw, ids)
}
\arguments{
\item{data_raw}{\code{data_raw}-object}

\item{ids}{Vector, specifying the columns in 'data_raw' that identify a decision maker uniquely}
}
\value{
vector, containing a unique identifier for each decision maker
}
\description{
Function that combines multiple identifier in 'data_raw' into a single identifier.
}
\keyword{internal}
