% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mappings.R
\name{make_mappings}
\alias{make_mappings}
\title{Create mapping matrices}
\usage{
make_mappings(mod, corr_re, error_struc)
}
\arguments{
\item{mod}{\code{mod}-object}

\item{corr_re}{Boolean, determining whether correlation for the random effects are estimated or not}

\item{error_struc}{Character, specifying the error structure}
}
\value{
\code{mod}-object
}
\description{
Function that creates mapping matrices Hb, fb, HO, fO, HL, fL if they do not exist in 'mod'.
These mappings define the structure of the Sigma matrix. Options are: \code{"fixed"} (\code{0.5*I}), \code{"ident"} (\code{sigma^2*I}), \code{"diag"}, \code{"full"}.
}
\keyword{internal}
