% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{makeAM}
\alias{makeAM}
\title{generate identity matrix of size 'alt' that adds a zero y'th row.}
\usage{
makeAM(y, alt)
}
\arguments{
\item{y}{integer; row to zero out}

\item{alt}{integer; numbre of alternatives}
}
\value{
matrix; altxalt
}
\description{
generate identity matrix of size 'alt' that adds a zero y'th row.
}
\keyword{internal}
