% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ME}
\alias{ME}
\title{Mendell-Elston approximation to multivariate Gaussian CDF.}
\usage{
ME(x, r)
}
\arguments{
\item{x}{nx1 vector of point to evaluate the CDF at.}

\item{r}{nxn correlation matrix.}
}
\value{
double; log of probability.
}
\description{
The function computes the  CDF for a multivariate Gaussian distribution approximated according to the method of Mendell-Elston.
}
\keyword{internal}
