% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ME_hess_new}
\alias{ME_hess_new}
\title{Hessian for Mendell-Elston approximation to multivariate Gaussian CDF.}
\usage{
ME_hess_new(x, r)
}
\arguments{
\item{x}{nx1 vector of point to evaluate the CDF at.}

\item{r}{nxn correlation matrix.}
}
\value{
Matrix; Hessian of log of probability.
}
\description{
The function computes the Hessian of the CDF for a multivariate Gaussian distribution approximated according to the method of Mendell-Elston.
}
\keyword{internal}
