% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AR_cl.R
\name{mod_AR_cl}
\alias{mod_AR_cl}
\title{R6 Object Representing Model Specifications}
\description{
A \code{mod_AR_cl} object contains the probit model specification for temporally dependent error processes.
The temporal correlation is introduced via an autoregressive system.
In the model class the lag length p is stored, as well as the initial conditions.
}
\section{Super class}{
\code{\link[Rprobit:mod_cl]{Rprobit::mod_cl}} -> \code{mod_AR_cl}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lag_length}}{lag length of the autoregression.}

\item{\code{stationary}}{indicates whether the stationary distribution shall be used to start the state process}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{tot_params}}{number of parameters for beta}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mod_AR_cl-set_lag_length}{\code{mod_AR_cl$set_lag_length()}}
\item \href{#method-mod_AR_cl-set_stationary}{\code{mod_AR_cl$set_stationary()}}
\item \href{#method-mod_AR_cl-print}{\code{mod_AR_cl$print()}}
\item \href{#method-mod_AR_cl-clone}{\code{mod_AR_cl$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="mod_cl" data-id="initialize"><a href='../../Rprobit/html/mod_cl.html#method-mod_cl-initialize'><code>Rprobit::mod_cl$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="mod_cl" data-id="set_alt"><a href='../../Rprobit/html/mod_cl.html#method-mod_cl-set_alt'><code>Rprobit::mod_cl$set_alt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="mod_cl" data-id="validate"><a href='../../Rprobit/html/mod_cl.html#method-mod_cl-validate'><code>Rprobit::mod_cl$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_AR_cl-set_lag_length"></a>}}
\if{latex}{\out{\hypertarget{method-mod_AR_cl-set_lag_length}{}}}
\subsection{Method \code{set_lag_length()}}{
Set dimension of state sequence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_AR_cl$set_lag_length(lag_length)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lag_length}}{An \code{integer}, the dimension of the state sequence}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_AR_cl-set_stationary"></a>}}
\if{latex}{\out{\hypertarget{method-mod_AR_cl-set_stationary}{}}}
\subsection{Method \code{set_stationary()}}{
Set stationary initial distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_AR_cl$set_stationary(stationary)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{stationary}}{A \code{Boolean}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_AR_cl-print"></a>}}
\if{latex}{\out{\hypertarget{method-mod_AR_cl-print}{}}}
\subsection{Method \code{print()}}{
prints the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_AR_cl$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_AR_cl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mod_AR_cl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_AR_cl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
