% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cl.R
\name{mod_cl}
\alias{mod_cl}
\title{R6 Object Representing Model Specifications}
\description{
A \code{mod_cl} object contains the probit model specification.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alt}}{number of alternatives}

\item{\code{ordered}}{\code{TRUE} for ordered choices and \code{FALSE} else}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{lthb}}{number of parameters for beta}

\item{\code{lthO}}{number of parameters for Omega}

\item{\code{lRE}}{number of random beta coefficients}

\item{\code{lthL}}{number of parameters for Sigma}

\item{\code{fb}}{vector of fixed parameters for beta}

\item{\code{Hb}}{matrix for dependence of beta on parameters}

\item{\code{fO}}{vector of fixed params for Omega}

\item{\code{HO}}{matrix for dependence of Omega parameters}

\item{\code{fL}}{vector of fixed params for Sigma}

\item{\code{HL}}{matrix for dependence of Sigma parameters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mod_cl-new}{\code{mod_cl$new()}}
\item \href{#method-mod_cl-validate}{\code{mod_cl$validate()}}
\item \href{#method-mod_cl-set_alt}{\code{mod_cl$set_alt()}}
\item \href{#method-mod_cl-print}{\code{mod_cl$print()}}
\item \href{#method-mod_cl-clone}{\code{mod_cl$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_cl-new"></a>}}
\if{latex}{\out{\hypertarget{method-mod_cl-new}{}}}
\subsection{Method \code{new()}}{
initialization function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_cl$new(
  Hb = matrix(0, 0, 0),
  fb = matrix(0, 0, 0),
  HO = matrix(0, 0, 0),
  fO = matrix(0, 0, 0),
  HL = matrix(0, 0, 0),
  fL = matrix(0, 0, 0),
  alt = 2,
  ordered = FALSE,
  validate = TRUE,
  check_identifiability = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Hb}}{matrix}

\item{\code{fb}}{vector}

\item{\code{HO}}{matrix}

\item{\code{fO}}{vector}

\item{\code{HL}}{matrix}

\item{\code{fL}}{vector}

\item{\code{alt}}{number of alternatives}

\item{\code{ordered}}{\code{TRUE} for ordered choices and \code{FALSE} else}

\item{\code{validate}}{\code{TRUE} to validate specification}

\item{\code{check_identifiability}}{\code{TRUE} to check for identifiability}

\item{\code{Tp}}{number of choices per decider}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_cl-validate"></a>}}
\if{latex}{\out{\hypertarget{method-mod_cl-validate}{}}}
\subsection{Method \code{validate()}}{
validates if parameters are conformable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_cl$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_cl-set_alt"></a>}}
\if{latex}{\out{\hypertarget{method-mod_cl-set_alt}{}}}
\subsection{Method \code{set_alt()}}{
Set number of alternatives.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_cl$set_alt(alt)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alt}}{An \code{integer}, the number of alternatives.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_cl-print"></a>}}
\if{latex}{\out{\hypertarget{method-mod_cl-print}{}}}
\subsection{Method \code{print()}}{
prints the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_cl$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_cl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mod_cl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_cl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
