% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_latclass_cl.R
\name{mod_latclass_cl}
\alias{mod_latclass_cl}
\title{R6 Object Representing Model Specifications}
\description{
A \code{mod_latclass_cl} object contains the probit model specification for latent class model.
}
\section{Super class}{
\code{\link[Rprobit:mod_cl]{Rprobit::mod_cl}} -> \code{mod_latclass_cl}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{num_class}}{number of classes}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{tot_params}}{number of parameters for beta}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mod_latclass_cl-set_num_class}{\code{mod_latclass_cl$set_num_class()}}
\item \href{#method-mod_latclass_cl-print}{\code{mod_latclass_cl$print()}}
\item \href{#method-mod_latclass_cl-clone}{\code{mod_latclass_cl$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="mod_cl" data-id="initialize"><a href='../../Rprobit/html/mod_cl.html#method-mod_cl-initialize'><code>Rprobit::mod_cl$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="mod_cl" data-id="set_alt"><a href='../../Rprobit/html/mod_cl.html#method-mod_cl-set_alt'><code>Rprobit::mod_cl$set_alt()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Rprobit" data-topic="mod_cl" data-id="validate"><a href='../../Rprobit/html/mod_cl.html#method-mod_cl-validate'><code>Rprobit::mod_cl$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_latclass_cl-set_num_class"></a>}}
\if{latex}{\out{\hypertarget{method-mod_latclass_cl-set_num_class}{}}}
\subsection{Method \code{set_num_class()}}{
Set number of classes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_latclass_cl$set_num_class(num_class)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_class}}{An \code{integer}, the number of classes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_latclass_cl-print"></a>}}
\if{latex}{\out{\hypertarget{method-mod_latclass_cl-print}{}}}
\subsection{Method \code{print()}}{
prints the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_latclass_cl$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mod_latclass_cl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-mod_latclass_cl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{mod_latclass_cl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
