% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlm_out_2_data_frame.R
\name{nlm_out_2_data_frame}
\alias{nlm_out_2_data_frame}
\title{Transform nlm_output.txt file to data frame}
\usage{
nlm_out_2_data_frame(
  file_name,
  save_in_file = FALSE,
  return_dataframe = TRUE,
  parameter_names = NULL
)
}
\arguments{
\item{file_name}{name of the .txt file in which the nlm output is saved}

\item{save_in_file}{either a boolean, indicating if the data frame should be saved in a .csv-file or a character string indicating the name of the file in which the data frame should be saved.}

\item{return_dataframe}{boolean, indicating if the data frame should be returned.}

\item{parameter_names}{optional character vector, specifying the parameter names.}
}
\value{
A data frame with iteration number, function value, parameter values and gradient values for each iteration of the nlm function.
}
\description{
Function that takes saved nlm output and extracts the information of the iterates and saves them in a data frame and if specified into a .csv file.
}
\keyword{internal}
