% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_data.R
\name{normalise_data}
\alias{normalise_data}
\title{Normalise data}
\usage{
normalise_data(data_raw, normalisation = FALSE, allvars = NULL, ASC = TRUE)
}
\arguments{
\item{data_raw}{\code{data_raw_cl}-object}

\item{normalisation}{\code{normalisation}-object. Either a list with normalisation specifications or a boolean.}

\item{allvars}{\code{allvars}-object}

\item{ASC}{boolean to check if ASCs are included in the model}
}
\value{
\code{data_raw_cl}-object with normalisation specifications as attribute.
}
\description{
Function that normalises data_raw regressor variables
}
\keyword{internal}
