% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{param_to_system_AR_R}
\alias{param_to_system_AR_R}
\title{converts parameters to state space system matrices corresponding to AR system and returns a list to R.}
\usage{
param_to_system_AR_R(param, lag, grad_bool, stationary)
}
\arguments{
\item{param}{parameter vector}

\item{lag}{integer; lag length.}

\item{grad_bool}{integer; if 0 the system is calculated, if non-zero the returned system is the derivative. The difference lies in entries fixed to one in the system which are zero for the derivative.}

\item{stationary}{boolean; if true, the state is started at its stationary distribution, else at zero.}
}
\value{
state_space_system as a list containing the system.
}
\description{
Computes the state space system with output dimension 1 and state dimension lag,
corresponding to the parameters param for the AR process. Passes the parameters to the C++ code.
}
