% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{param_to_system_R}
\alias{param_to_system_R}
\title{converts parameters to state space system matrices and returns a list to R.}
\usage{
param_to_system_R(param, s, n, grad_bool, stationary)
}
\arguments{
\item{param}{parameter vector}

\item{s}{integer; output dimension}

\item{n}{integer; state dimension}

\item{grad_bool}{integer; if 0 the system is calculated, if non-zero the returned system is the derivative. The difference lies in entries fixed to one in the system which are zero for the derivative.}

\item{stationary}{boolean; if true, the state is started at its stationary distribution, else at zero.}
}
\value{
state_space_system as a list containing the system.
}
\description{
Computes the state space system with output dimension s and state dimension n,
corresponding to the parameters param. Passes the parameters to the C++ code.
}
