% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Rprobit_cl.R
\name{plot.Rprobit_cl}
\alias{plot.Rprobit_cl}
\title{Plot the coefficient distribution function
(makes sense, if random coefficients or latent classes exist)}
\usage{
\method{plot}{Rprobit_cl}(x, ...)
}
\arguments{
\item{x}{Rprobit_cl object containing the model.}

\item{...}{passed on to the plot function of the Rprobit_cl class.}
}
\value{
A plot.
}
\description{
This function provides a plot for the
}
