% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Rprobit.R
\name{plot_Rprobit}
\alias{plot_Rprobit}
\title{Plot the coefficient distribution function
(makes sense, if random coefficients or latent classes exist)}
\usage{
plot_Rprobit(object, dims = 1, cdf = FALSE, margin = 0.1)
}
\arguments{
\item{object}{An \code{\link{Rprobit_cl}} object.}

\item{dims}{vector of integers indicating the coordinates to be plotted. Only the first two integers are considered}

\item{cdf}{Boolean, indicating whether the cdf (TRUE) or the pdf (FALSE; default) should be plotted.}

\item{margin}{real; factor (0.1 = 10\%; minimum default margin) to enlarge plotting area.}
}
\value{
A plot.
}
\description{
This function provides a plot for the
}
