% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pred_probit_approx}
\alias{pred_probit_approx}
\title{Choice probabilities for probit models}
\usage{
pred_probit_approx(theta, data, mod, control)
}
\arguments{
\item{theta}{parameter vector}

\item{data}{data list}

\item{mod}{A \code{\link{mod_cl}} object.}

\item{control}{Controls for the probit estimation.}
}
\value{
A vector, containing the negative log-likelihood with attributes containing the gradient and (if specified in the controls) the Hessian.
}
\description{
Computes the approximate choice porbabilities for the probit case.
}
\keyword{internal}
