% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pred_probit_ordered_approx}
\alias{pred_probit_ordered_approx}
\title{Choice probabilities for ordered probit models}
\usage{
pred_probit_ordered_approx(theta, Xn, yn, mod)
}
\arguments{
\item{theta}{parameter vector}

\item{Xn}{matrix of regressors}

\item{yn}{vector of responses}

\item{mod}{A \code{\link{mod_cl}} object.}
}
\value{
A matrix, containing the predicted choice probabilities for each choice marginally (ignoring correlations)
}
\description{
Computes the approximate choice probabilities for the ordered probit case.
}
