% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.Rprobit_cl.R
\name{predict.Rprobit_cl}
\alias{predict.Rprobit_cl}
\title{Prediction Based on a Fitted Model}
\usage{
\method{predict}{Rprobit_cl}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An \code{\link{Rprobit_cl}} object.}

\item{newdata}{Either \code{\link{data_raw_cl}} object. Or a data.frame of the same structure as the ones on which
the \code{\link{Rprobit_cl}} \code{object} is based upon.}

\item{...}{Not used.}
}
\value{
A \code{list} of matrices with information on the estimated model.
}
\description{
This function calculates the in-sample predictions for a fitted model.
}
