% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_Rprobit.R
\name{predict_Rprobit}
\alias{predict_Rprobit}
\title{Calculate Choice Predictions}
\usage{
predict_Rprobit(Rprobit_obj, data_new = NULL, all_pred = FALSE)
}
\arguments{
\item{Rprobit_obj}{An \code{\link{Rprobit_cl}} object.}

\item{data_new}{Optionally new data for prediction. Can be
\itemize{
\item a \code{\link{data_cl}} object
\item or a \code{\link{data_raw_cl}} object.
By default, \code{data_new = NULL}.
}}

\item{all_pred}{A \code{logical}, indicating whether mean predicted probabilities or the
matrix of all predicted probabilities should be returned.}
}
\value{
A \code{matrix}, true choice percentages in first and averages of predictions
in second row
}
\description{
This function calculates and returns true choice percentages and averages of
choice prediction.
}
