% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_est.R
\name{print_est}
\alias{print_est}
\title{Formatted printing of a matrix}
\usage{
print_est(est_sd)
}
\arguments{
\item{est_sd}{matrix}
}
\value{
inline written text
}
\description{
Function that prints a formatted matrix.
}
\keyword{internal}
