% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Rprobit_cl.R
\name{print_est_sd}
\alias{print_est_sd}
\title{Print Estimated and Standard Deviations}
\usage{
print_est_sd(est_sd, row_names = NULL, col_names = NULL)
}
\arguments{
\item{est_sd}{A \code{matrix} of dimension \code{m} times \code{2n}, where the first
\code{m} columns are estimates and the
last \code{m} columns are corresponding standard deviations.}

\item{row_names}{A \code{character} vector of length \code{m} with row names.
Can be \code{NULL} (default) for no row names.}

\item{col_names}{A \code{character} vector of length \code{n} with column names.
Can be \code{NULL} (default) for no column names.}
}
\value{
No return value, prints a \code{matrix}.
}
\description{
This helper function prints a formatted version of a matrix with estimates
and standard deviations.
}
\examples{
\dontrun{
est_sd <- matrix(1:12, nrow = 3, ncol = 4)
print_est_sd(est_sd, row_names = LETTERS[1:3], col_names = LETTERS[4:5])
}

}
\keyword{internal}
\keyword{utils}
