% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prob_ordered}
\alias{prob_ordered}
\title{Probability to multivariate ordered probit model.}
\usage{
prob_ordered(xb, y_n, Lambda, alt, dtauk, approx_method)
}
\arguments{
\item{xb}{nx1 vector of systematic part}

\item{y_n}{nx1 integer vector of choices.}

\item{Lambda}{Covariance matrix for random utilities including the random coefficients.}

\item{alt}{integer; number of choice alternatives.}

\item{dtauk}{(alt-1) real vector of increments of category limits obtained as the cumulative sum of the vector.}

\item{approx_method}{string: approximation method, either "SJ" or "ME"}
}
\value{
double; log of probability.
}
\description{
The function computes the probability of a number of choices in repeated ordered probit choices.
}
\keyword{internal}
