% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prob_ordered_1}
\alias{prob_ordered_1}
\title{calculates the probability of one choice in an ordered probit model}
\usage{
prob_ordered_1(xb, y1, Lambda, alt, dtauk)
}
\arguments{
\item{xb}{real; systematic utility}

\item{y1}{integer; choice}

\item{Lambda}{real; variance}

\item{alt}{integer; number of alternatives}

\item{dtauk}{vector of category limits.}
}
\value{
matrix
}
\description{
calculates the probability of one choice in an ordered probit model
}
\keyword{internal}
