% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prob_ordered_2}
\alias{prob_ordered_2}
\title{calculates the probability of pair of choices in an ordered probit model}
\usage{
prob_ordered_2(xb, y, Lambda, alt, dtauk)
}
\arguments{
\item{xb}{vector of systematic utilities}

\item{y}{vector of choices}

\item{Lambda}{correlation matrix}

\item{alt}{integer; number of alternatives}

\item{dtauk}{vector of category limits.}
}
\value{
matrix
}
\description{
calculates the probability of pair of choices in an ordered probit model
}
\keyword{internal}
