% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prob_ordered_CML}
\alias{prob_ordered_CML}
\title{calculates the CML of a number of choices in an ordered probit model}
\usage{
prob_ordered_CML(xb, y, Lambda, alt, dtauk, cml_pair_type)
}
\arguments{
\item{xb}{vector of systematic utilities}

\item{y}{vector of choices}

\item{Lambda}{correlation matrix}

\item{alt}{integer; number of alternatives}

\item{dtauk}{vector of category limits.}

\item{cml_pair_type}{integer; indicating pair type structure}
}
\value{
matrix
}
\description{
calculates the CML of a number of choices in an ordered probit model
}
\keyword{internal}
