% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_formula.R
\name{read_formula}
\alias{read_formula}
\title{Read formula}
\usage{
read_formula(form)
}
\arguments{
\item{form}{An object of class \code{formula} of the form
\code{choice ~ Type 1 | Type 2 | Type 3}, where:
\itemize{
\item \code{choice} indicates the choice variable
\item \code{Type 1} regressors are varying over alternatives and enter the model as is
\item \code{Type 2} regressors are constant over alternatives and enter the model with alternative varying coefficients
(setting the coefficient for norming alternative to zero)
\item \code{Type 3} regressors are alternative varying and enter the utilities with alternatives varying coefficients
}}
}
\value{
A list containing the following elements:
\item{allvars}{\code{allvars}-object (list of three entries listing the Type 1, 2 and 3 variables)}
\item{ASC}{Boolean, indicating whether ASCs are estimated or not}
}
\description{
Function that reads out different types of covariates from formula object 'form'.
}
\keyword{internal}
