% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redraw_choices_from_data.r
\name{redraw_choices_from_data}
\alias{redraw_choices_from_data}
\title{Redraw data from Rprobit_cl object}
\usage{
redraw_choices_from_data(Rprobit_obj, seed = 1, thetas = NULL)
}
\arguments{
\item{Rprobit_obj}{An \code{\link{Rprobit_cl}} object.}

\item{seed}{Seed for the simulation of choice data}

\item{thetas}{real matrix, each row corresponds to one value of the parameters. When number of persons matches, each person gets one parameter. Else parameter vectors are drawn randomly.}
}
\value{
list an updated \code{\link{Rprobit_cl}} object.
}
\description{
This function simulates new choices from data stored in an Rprobit_cl object, which contains full information on regressors and model and parameters.
}
