% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_LC_model.R
\name{reorder_LC_model}
\alias{reorder_LC_model}
\title{Build a covariance matrix and calculates the standard deviations}
\usage{
reorder_LC_model(Rprobit_obj)
}
\arguments{
\item{Rprobit_obj}{A \code{\link{Rprobit_cl}} object.}
}
\value{
The reordered \code{\link{Rprobit_cl}} object.
}
\description{
Function that builds a covariance matrix and calculates the standard deviations using the Delta method.
}
\keyword{internal}
