% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.Rprobit_cl.R
\name{residuals.Rprobit_cl}
\alias{residuals.Rprobit_cl}
\title{Provide Residuals for a Fitted Model}
\usage{
\method{residuals}{Rprobit_cl}(object, ...)
}
\arguments{
\item{object}{An \code{\link{Rprobit_cl}} object.}

\item{...}{Not used.}
}
\value{
A \code{matrix}, the residuals.
}
\description{
This function provides the residuals for a fitted model.
}
