% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rprobit-package.R
\docType{package}
\name{Rprobit-package}
\alias{Rprobit}
\alias{Rprobit-package}
\title{Rprobit: Estimation of Multinomial Probit Models}
\description{
Estimation of multinomial probit models based on panel data. It uses the full likelihood as well as the MaCML approach. The multinomial Gaussian CDF is approximated using the Solow-Joe, Mendell-Elston or the TVBS method of Chandra Bhat.
}
\author{
\strong{Maintainer}: Dietmar Bauer \email{dietmar.bauer@uni-bielefeld.de}

Authors:
\itemize{
  \item Manuel Batram
  \item Sebastian Büscher \email{sebastian.buescher@uni-bielefeld.de}
  \item Lennart Oelschläger \email{lennart.oelschlaeger@uni-bielefeld.de}
}

}
\keyword{internal}
