% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rprobit.R
\docType{package}
\name{Rprobit}
\alias{Rprobit}
\title{Rprobit: An R package for the estimation of multinomial probit models}
\description{
Rprobit provides the estimation of multinomial probit models based
on panel data. It uses the full likelihood as well as the MaCML approach.
The multinomial Gaussian CDF is approximated using the Solow-Joe or the TVBS
method of Chandra Bhat.
}
\keyword{internal}
