% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rprobit_cl.R, R/summary.Rprobit_cl.R
\name{Rprobit_cl}
\alias{Rprobit_cl}
\alias{summary.Rprobit_cl}
\title{R6 Object Representing a Probit Model}
\usage{
\method{summary}{Rprobit_cl}(object, denormalise = TRUE, ...)
}
\arguments{
\item{object}{An \code{\link{Rprobit_cl}} object.}

\item{denormalise}{A \code{logical}, decides if summary should be given with respect to original
variables (\code{denormalise = TRUE}) or normalised variables
(\code{denormalise = FALSE}).
By default, \code{denormalise = TRUE}.}

\item{...}{Not used.}
}
\value{
A \code{summary.Rprobit_cl} object, which is a \code{list} that contains
the following information on the estimated model:
\itemize{
\item \code{N}, the number of decision makers
\item \code{alt}, the number of alternatives
\item \code{Tp}, the vector of number of choice occasions
\item \code{omega_present}, a \code{logical}, indicating whether an Omega matrix exists
\item \code{vars}, a \code{character} vector of variable names
\item \code{alt_names}, a \code{character} vector of alternative names
\item \code{ordered}, a \code{logical}, indicating whether the alternatives are ordered
\item \code{par}, a \code{list} of parameter estimates
\item \code{prediction_summary}, a \code{matrix} of prediction details
\item \code{ll}, a \code{numeric}, the log-likelihood value
}
}
\description{
An \code{Rprobit_cl} object contains the probit model specification.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{control}}{An \code{\link{control_cl}} object which stores information for estimation.}

\item{\code{data}}{A \code{\link{data_cl}} object.}

\item{\code{data_raw}}{A \code{\link{data_raw_cl}} object.}

\item{\code{form}}{A \code{formula} object.}

\item{\code{re}}{A \code{character} vector containing the variable names for random effects.}

\item{\code{mod}}{A \code{\link{mod_cl}} object.}

\item{\code{vars}}{A \code{character} vector containing variable names.}

\item{\code{alt_names}}{A \code{character} vector containing alternative names.}

\item{\code{theta}}{A \code{numeric} parameter vector.}

\item{\code{theta_0}}{A \code{numeric} true parameter vector (for simulated data only).}

\item{\code{ll}}{A \code{numeric}, the log likelihood value.}

\item{\code{H}}{A \code{matrix}, the Hessian at estimate.}

\item{\code{J}}{A variance \code{matrix} of score vector.}

\item{\code{grad}}{A gradient \code{vector} at estimate.}

\item{\code{fit}}{A \code{character}, the approximation method.}

\item{\code{vv}}{A variance \code{matrix} computed using the Hessian.}

\item{\code{vv2}}{A variance \code{matrix} of different variance estimate computed using
the approximated Hessian.}

\item{\code{info}}{A \code{list} containing information on fitting process.}

\item{\code{gradEL}}{list containing gradient contribution of each individual}

\item{\code{HessEL}}{list containing Hessian contribution of each individual}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Rprobit_cl-new}{\code{Rprobit_cl$new()}}
\item \href{#method-Rprobit_cl-set_data}{\code{Rprobit_cl$set_data()}}
\item \href{#method-Rprobit_cl-set_data_raw}{\code{Rprobit_cl$set_data_raw()}}
\item \href{#method-Rprobit_cl-set_form}{\code{Rprobit_cl$set_form()}}
\item \href{#method-Rprobit_cl-set_re}{\code{Rprobit_cl$set_re()}}
\item \href{#method-Rprobit_cl-set_mod}{\code{Rprobit_cl$set_mod()}}
\item \href{#method-Rprobit_cl-set_vars}{\code{Rprobit_cl$set_vars()}}
\item \href{#method-Rprobit_cl-set_alt_names}{\code{Rprobit_cl$set_alt_names()}}
\item \href{#method-Rprobit_cl-set_theta}{\code{Rprobit_cl$set_theta()}}
\item \href{#method-Rprobit_cl-set_theta_0}{\code{Rprobit_cl$set_theta_0()}}
\item \href{#method-Rprobit_cl-set_ll}{\code{Rprobit_cl$set_ll()}}
\item \href{#method-Rprobit_cl-set_H}{\code{Rprobit_cl$set_H()}}
\item \href{#method-Rprobit_cl-set_J}{\code{Rprobit_cl$set_J()}}
\item \href{#method-Rprobit_cl-set_grad}{\code{Rprobit_cl$set_grad()}}
\item \href{#method-Rprobit_cl-set_fit}{\code{Rprobit_cl$set_fit()}}
\item \href{#method-Rprobit_cl-set_vv}{\code{Rprobit_cl$set_vv()}}
\item \href{#method-Rprobit_cl-set_vv2}{\code{Rprobit_cl$set_vv2()}}
\item \href{#method-Rprobit_cl-set_control}{\code{Rprobit_cl$set_control()}}
\item \href{#method-Rprobit_cl-set_info}{\code{Rprobit_cl$set_info()}}
\item \href{#method-Rprobit_cl-set_gradEL}{\code{Rprobit_cl$set_gradEL()}}
\item \href{#method-Rprobit_cl-set_HessEL}{\code{Rprobit_cl$set_HessEL()}}
\item \href{#method-Rprobit_cl-print}{\code{Rprobit_cl$print()}}
\item \href{#method-Rprobit_cl-coef}{\code{Rprobit_cl$coef()}}
\item \href{#method-Rprobit_cl-plot}{\code{Rprobit_cl$plot()}}
\item \href{#method-Rprobit_cl-predict}{\code{Rprobit_cl$predict()}}
\item \href{#method-Rprobit_cl-residuals}{\code{Rprobit_cl$residuals()}}
\item \href{#method-Rprobit_cl-AIC}{\code{Rprobit_cl$AIC()}}
\item \href{#method-Rprobit_cl-BIC}{\code{Rprobit_cl$BIC()}}
\item \href{#method-Rprobit_cl-conf_mat}{\code{Rprobit_cl$conf_mat()}}
\item \href{#method-Rprobit_cl-clone}{\code{Rprobit_cl$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-new"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-new}{}}}
\subsection{Method \code{new()}}{
initialization function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$new(
  data = NULL,
  data_raw = NULL,
  form = NULL,
  re = NULL,
  mod = NULL,
  vars = NULL,
  alt_names = NULL,
  theta = NULL,
  theta_0 = NULL,
  ll = NA,
  H = NA,
  J = NA,
  grad = NA,
  fit = NA,
  vv = NA,
  vv2 = NA,
  control = NULL,
  info = NULL,
  gradEL = NULL,
  HessEL = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{object of class \link{data_cl}}

\item{\code{data_raw}}{object of class \link{data_raw_cl}}

\item{\code{form}}{\link{formula} object}

\item{\code{re}}{list of strings containing the variable names for random effects}

\item{\code{mod}}{object of class \link{mod_cl}}

\item{\code{vars}}{list of strings containing variable names}

\item{\code{alt_names}}{list of strings of names for alternatives}

\item{\code{theta}}{parameter vector}

\item{\code{theta_0}}{true parameter vector for simulated data}

\item{\code{ll}}{log likelihood value}

\item{\code{H}}{matrix of Hessian at parameter estimate}

\item{\code{J}}{matrix of variance of score vector}

\item{\code{grad}}{gradient vector at estimate}

\item{\code{fit}}{approximation method}

\item{\code{vv}}{matrix of variance computed using the Hessian}

\item{\code{vv2}}{matrix of different variance estimate computed using the approximated Hessian}

\item{\code{control}}{object of class \link{control_cl}. Stores all information for estimation.}

\item{\code{info}}{list containing information on fitting process}

\item{\code{gradEL}}{list containing gradient contribution of each individual}

\item{\code{HessEL}}{list containing Hessian contribution of each individual}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_data}{}}}
\subsection{Method \code{set_data()}}{
Set data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_data(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_data_raw"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_data_raw}{}}}
\subsection{Method \code{set_data_raw()}}{
Set data_raw
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_data_raw(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{New data_raw}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_form"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_form}{}}}
\subsection{Method \code{set_form()}}{
Set form
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_form(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{form}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_re"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_re}{}}}
\subsection{Method \code{set_re()}}{
Set re
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_re(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{list of strings containing names of random coefficients}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_mod"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_mod}{}}}
\subsection{Method \code{set_mod()}}{
Set mod
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_mod(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{object of class 'mod_cl'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_vars"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_vars}{}}}
\subsection{Method \code{set_vars()}}{
Set vars
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_vars(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{list of strings containing names of variables}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_alt_names"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_alt_names}{}}}
\subsection{Method \code{set_alt_names()}}{
Set alt_names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_alt_names(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{list of strings containing names of alternatives}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_theta"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_theta}{}}}
\subsection{Method \code{set_theta()}}{
Set theta
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_theta(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{numeric vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_theta_0"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_theta_0}{}}}
\subsection{Method \code{set_theta_0()}}{
Set theta_0
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_theta_0(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{numeric vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_ll"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_ll}{}}}
\subsection{Method \code{set_ll()}}{
Set ll
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_ll(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{double}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_H"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_H}{}}}
\subsection{Method \code{set_H()}}{
Set H
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_H(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{Hessian matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_J"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_J}{}}}
\subsection{Method \code{set_J()}}{
Set J
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_J(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{varianc matrix of score}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_grad"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_grad}{}}}
\subsection{Method \code{set_grad()}}{
Set grad
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_grad(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{numeric vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_fit"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_fit}{}}}
\subsection{Method \code{set_fit()}}{
Set fit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_fit(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{fit}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_vv"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_vv}{}}}
\subsection{Method \code{set_vv()}}{
Set vv
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_vv(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{variance matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_vv2"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_vv2}{}}}
\subsection{Method \code{set_vv2()}}{
Set vv2
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_vv2(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{approximate variance matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_control"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_control}{}}}
\subsection{Method \code{set_control()}}{
Set control
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_control(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{control object of class \link{control_cl}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_info"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_info}{}}}
\subsection{Method \code{set_info()}}{
Set info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_info(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{info}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_gradEL"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_gradEL}{}}}
\subsection{Method \code{set_gradEL()}}{
Set gradEL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_gradEL(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-set_HessEL"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-set_HessEL}{}}}
\subsection{Method \code{set_HessEL()}}{
Set HessEL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$set_HessEL(val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-print"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-print}{}}}
\subsection{Method \code{print()}}{
print object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-coef"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-coef}{}}}
\subsection{Method \code{coef()}}{
return coefficients for beta parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$coef()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
beta
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-plot}{}}}
\subsection{Method \code{plot()}}{
provide plots for model (useful for latent class models)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$plot(dims = 1, cdf = FALSE, margin = 0.1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dims}}{vector of integers indicating the coordinates to be plotted. Only the first two integers are considered}

\item{\code{cdf}}{Boolean, indicating whether the cdf (TRUE) or the pdf (FALSE; default) should be plotted.}

\item{\code{margin}}{real; factor (0.1 = 10\%; minimum default margin) to enlarge plotting area.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-predict}{}}}
\subsection{Method \code{predict()}}{
provide predictions for model contained in object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$predict(newdata = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{either a \code{\link{data_raw_cl}} object or a data frame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-residuals"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-residuals}{}}}
\subsection{Method \code{residuals()}}{
provide residuals for model contained in object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$residuals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-AIC"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-AIC}{}}}
\subsection{Method \code{AIC()}}{
provide AIC value for model contained in object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$AIC()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-BIC"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-BIC}{}}}
\subsection{Method \code{BIC()}}{
provide BIC value for model contained in object. Sample size equal to N*T in balanced case
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$BIC()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-conf_mat"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-conf_mat}{}}}
\subsection{Method \code{conf_mat()}}{
provide confusion matrix for model contained in object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$conf_mat()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rprobit_cl-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Rprobit_cl-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rprobit_cl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
