% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rprobit_cl_from_modMACML.R
\name{Rprobit_cl_from_modMACML}
\alias{Rprobit_cl_from_modMACML}
\title{Create Rprobit-object from legacy mod_macml object}
\usage{
Rprobit_cl_from_modMACML(mod_object)
}
\arguments{
\item{mod_object}{\code{mod_macml}-object}
}
\value{
An \code{\link{Rprobit_cl}} object.
}
\description{
This function creates an \code{\link{Rprobit_cl}} object from a legacy mod_macml object from previous package iterations.
}
\keyword{internal}
