% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mini_example.R
\name{run_mini_example}
\alias{run_mini_example}
\title{Perform mini simulation examples}
\usage{
run_mini_example(
  form,
  mod,
  re,
  approx_method = "SJ",
  show_theta = FALSE,
  seed = NULL,
  at_true = FALSE,
  runs = 1,
  probit = TRUE,
  nCores = 1,
  print.level = 0,
  normalize = FALSE,
  control_simul = NULL,
  verbose = TRUE,
  compare_time = FALSE,
  theta = NULL,
  show_ll = FALSE,
  cml_pair_type = 0
)
}
\arguments{
\item{form}{A \code{form} object.}

\item{mod}{A \code{\link{mod_cl}} object.}

\item{re}{A \code{character} vector of variable names which are considered to have
random effects.}

\item{approx_method}{A \code{character}. Can be (one or more of)
\itemize{
\item \code{"SJ"} (default),
\item \code{"TVBS"},
\item \code{"TVBSv2"},
\item \code{"ME"}.
}}

\item{show_theta}{boolean, if true, compare theta vector, if false (default) compare entries of b, Omega and Sigma}

\item{seed}{set a seed for the initial values and the simulated data, no seed per default}

\item{at_true}{if \code{TRUE}, starts optimization from true parameter}

\item{runs}{number of numerical optimizations at different starting values, default \code{1}}

\item{probit}{if \code{TRUE} (default) estimates a probit model, else a CML-model}

\item{nCores}{number of cores used for estimation}

\item{print.level}{Passed on to \code{nlm}, default \code{0}}

\item{normalize}{boolean indicating whether probabilities are normalize to sum to 1, passed on to \code{nlm}, default \code{FALSE}}

\item{control_simul}{list containing at least field Tp controlling the simulation.}

\item{verbose}{boolean indicating whether to print on-screen information}

\item{compare_time}{boolean indicating whether to save estimation time}

\item{theta}{parameter vector}

\item{show_ll}{boolean indicating whether to save the log-likelihood}

\item{cml_pair_type}{integer specifying the CML to use (0: full pairwise, 1: adjacent lopped, 2: adjacent)
By default \code{cml_pair_type = 0} is used}
}
\value{
no return value; on-screen information
}
\description{
This function performs mini simulation examples.
}
\keyword{internal}
