% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_Rprobit.R
\name{setup_Rprobit}
\alias{setup_Rprobit}
\title{Create Rprobit-object}
\usage{
setup_Rprobit(
  form,
  data_raw = NULL,
  mod = NULL,
  norm_alt = 1,
  ids = NULL,
  re = NULL,
  corr_re = FALSE,
  error_struc = "full",
  seed = NULL,
  theta_0 = NULL,
  name = NULL,
  description = NULL,
  control = NULL,
  avail = NULL,
  info = NULL,
  alt_names = NULL
)
}
\arguments{
\item{form}{An object of class \code{formula} of the form
\code{choice ~ Type 1 | Type 2 | Type 3}, where:
\itemize{
\item \code{choice} indicates the choice variable
\item \code{Type 1} regressors are varying over alternatives and enter the model as is
\item \code{Type 2} regressors are constant over alternatives and enter the model with alternative varying coefficients
(setting the coefficient for norming alternative to zero)
\item \code{Type 3} regressors are alternative varying and enter the utilities with alternatives varying coefficients
}}

\item{data_raw}{\code{data_raw}-object}

\item{mod}{A \code{\link{mod_cl}}, \code{\link{mod_latclass_cl}} or \code{\link{mod_StSp_cl}} object.}

\item{norm_alt}{An \code{integer} that labels the alternative which's utility is set to zero
(for normalization).}

\item{ids}{A \code{character} vector, specifying the columns that identify a decision
maker uniquely.}

\item{re}{A \code{character} vector of variable names which are considered to have
random effects.}

\item{corr_re}{A \code{logical}, if \code{TRUE} a full covariance matrix Omega will be
setup, if \code{FALSE} (default) only a diagonal one.}

\item{error_struc}{A \code{character}, specifying the error structure, one of:
\itemize{
\item \code{"diag"} diagonal matrix Sigma
\item \code{"fixed"} fixed matrix Sigma
\item \code{"ident"} Sigma proportional to identity matrix.
\item \code{"full"} (default)
}}

\item{seed}{Seed for the simulation of choice data.
No seed set per default.}

\item{theta_0}{A \code{numeric} vector, the true parameter vector for simulated data.
It is drawn randomly if not supplied.}

\item{name}{A \code{character}, a name for the model.}

\item{description}{A \code{character}, a description for the model.}

\item{control}{A \code{list}, contains control variables for estimation procedure.}

\item{avail}{A \code{list}, contains information on the availability for each alternative
and each choice occasion.}

\item{info}{A \code{list}, contains information on the model and estimation performance.}

\item{alt_names}{A \code{character} vector, containing the labels for the choices.}
}
\value{
An \code{\link{Rprobit_cl}} object.
}
\description{
This function creates an \code{\link{Rprobit_cl}} object.
}
