% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_raw.R
\name{sim_data_raw}
\alias{sim_data_raw}
\title{Simulate data_raw}
\usage{
sim_data_raw(
  ASC,
  allvars,
  re,
  choice,
  theta,
  mod,
  seed,
  control_simulation = NULL
)
}
\arguments{
\item{ASC}{Boolean, indicating whether ASCs are estimated or not}

\item{allvars}{\code{allvars}-object}

\item{re}{Character vector of variable names which are considered to have random effects}

\item{choice}{Character, name of dependent variable}

\item{theta}{\code{theta}-object}

\item{mod}{A \code{\link{mod_cl}} or \code{\link{mod_latclass_cl}} object.}

\item{seed}{Seed for the simulation of choice data}

\item{control_simulation}{list to fine-tune the simulated data. In case auto correlated errors should be simulated, control_simulation$simulate_ar1_rho determines auto correlation coefficient/matrix}
}
\value{
list of \code{data_raw}, \code{data} and \code{theta}
}
\description{
This function simulates choice data.
}
\keyword{internal}
