% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SJ}
\alias{SJ}
\title{Solow-Joe approximation to multivariate Gaussian CDF.}
\usage{
SJ(x, r)
}
\arguments{
\item{x}{nx1 vector of point to evaluate the CDF at.}

\item{r}{nxn correlation matrix.}
}
\value{
double; log of probability.
}
\description{
The function computes the  CDF for a multivariate Gaussian distribution according to the method of Solow-Joe.
}
