% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SJ_hess_new}
\alias{SJ_hess_new}
\title{Hessian of Solow-Joe approximation to multivariate Gaussian CDF.}
\usage{
SJ_hess_new(x, r)
}
\arguments{
\item{x}{nx1 vector of point to evaluate the CDF at.}

\item{r}{nxn correlation matrix.}
}
\value{
matrix; Hessian of log of probability.
}
\description{
The function computes the Hessian of the  CDF for a multivariate Gaussian distribution according to the method of Solow-Joe.
}
\keyword{internal}
