% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_Lyapunov_equation}
\alias{solve_Lyapunov_equation}
\title{Solution to the Lyapunov equation (assumes A is stable)}
\usage{
solve_Lyapunov_equation(A, vQ)
}
\arguments{
\item{A}{Eigen::MatrixXd; matrix describing the dynamics.}

\item{vQ}{Eigen::VectorXd; vectorized version of the matrix Q.}
}
\value{
Eigen::MatrixXd; solution matrix.
}
\description{
Computes the solution to the Lyapunov equation P = APA' + Q
}
\keyword{internal}
