% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subset}
\alias{subset}
\title{take out rows and columns}
\usage{
subset(Gam, vec)
}
\arguments{
\item{Gam}{matrix.}

\item{vec}{vector indices.}
}
\value{
a submatrix.
}
\description{
Extracts the submatrix Gamma(vec,vec).
}
\keyword{internal}
