% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substract_choice_regressor_from_data.R
\name{substract_choice_regressor_from_data}
\alias{substract_choice_regressor_from_data}
\title{Transform data for estimation}
\usage{
substract_choice_regressor_from_data(data)
}
\arguments{
\item{data}{\code{data}-object}
}
\value{
transformed \code{data}-object \code{"data_tr"}
}
\description{
Function that transform 'data' by subtracting the regressor of the chosen alternative from the regressor matrix.
}
\keyword{internal}
