% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_lin_hypothesis.R
\name{test_lin_hypothesis}
\alias{test_lin_hypothesis}
\title{Test linear hypotheses on the parameter vector using empirical likelihood and adjusted likelihood ratios.}
\usage{
test_lin_hypothesis(Rprobit_obj, hypothesis)
}
\arguments{
\item{Rprobit_obj}{A \code{\link{Rprobit_cl}} object.}

\item{hypothesis}{Either a \code{\link{hypothesis_cl}} object or a vector for H_0: theta = theta_0.}
}
\value{
A real matrix where each row has two entries corresponding to the unrestricted and the restricted estimate of
\itemize{
\item CML ratio
\item empirical likelihood ratio
\item AIC
\item BIC
}
}
\description{
This function uses the estimate without restrictions to obtain an estimate under the restricted CML.
Subsequently different tests for a linear hypothesis on the parameter vector is performed
using the empirical likelihood ratio as well as adjustments to the CML ratio.
}
