% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TVBS_grad}
\alias{TVBS_grad}
\title{TVBS approximation to multivariate Gaussian CDF: calculates CDF and gradient.}
\usage{
TVBS_grad(x_norm, Cor_mat, log_out = 0L)
}
\arguments{
\item{x_norm}{nx1 vector of point to evaluate the CDF at.}

\item{Cor_mat}{nxn correlation matrix.}

\item{log_out}{integer; probability or log of probability?}
}
\value{
vector; gradient of log probability, double; (log of) probability.
}
\description{
The function computes the  CDF and the corresponding gradient for a multivariate Gaussian distribution according to the method of Chandra Bhat (TVBS).
}
\keyword{internal}
