% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TVBS_hess_new}
\alias{TVBS_hess_new}
\title{TVBS approximation to multivariate Gaussian CDF: calculates Hessian of log CDF}
\usage{
TVBS_hess_new(x_norm, Cor_mat)
}
\arguments{
\item{x_norm}{nx1 vector of point to evaluate the CDF at.}

\item{Cor_mat}{nxn correlation matrix.}
}
\value{
matrix; Hessian of log probability.
}
\description{
The function computes the Hessian of the log of the CDF and the corresponding gradient for a multivariate Gaussian distribution according to the method of Chandra Bhat (TVBS).
}
\keyword{internal}
