% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TVBS_pmvnorm_cpp}
\alias{TVBS_pmvnorm_cpp}
\title{multivariate Gaussian CDF: calculates CDF up to four dimensions}
\usage{
TVBS_pmvnorm_cpp(upper, mu, Sigma)
}
\arguments{
\item{upper}{nx1 vector of points where to evaluate}

\item{mu}{nx1 vector; expectation}

\item{Sigma}{nxn correlation matrix.}
}
\value{
vector; gradient of log probability, double; (log of) probability.
}
\description{
The function computes the  Gaussian CDF for 1, 2, 3 and four dimensional problems.
}
\keyword{internal}
