% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vechor}
\alias{vechor}
\title{vech indices}
\usage{
vechor(k)
}
\arguments{
\item{k}{integer; dimension of matrix.}
}
\value{
matrix
}
\description{
Returns indices of vech(X) w.r.t to the entries of X WITHOUT THE DIAGONAL
}
\keyword{internal}
