% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vechor_diag}
\alias{vechor_diag}
\title{Returns indices of vech(X) w.r.t to the entries of X}
\usage{
vechor_diag(k)
}
\arguments{
\item{k}{integer; dimension of matrix.}
}
\value{
matrix
}
\description{
Returns indices of vech(X) w.r.t to the entries of X
}
\keyword{internal}
