% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vectorize}
\alias{vectorize}
\title{vectorize a matrix columnwise}
\usage{
vectorize(Q)
}
\arguments{
\item{Q}{Eigen::MatrixXd; real matrix.}
}
\value{
Eigen::VectorXd; double vector.
}
\description{
Returns the vector of the columnwise vectorized rectangular matrix
}
\keyword{internal}
