% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xplus3.R
\name{Xplus3}
\alias{Xplus3}
\title{evaluate (x)_+^3 for splines}
\usage{
Xplus3(x)
}
\arguments{
\item{x}{A real value}
}
\value{
A real value.
}
\description{
This function is a helper function used for the evaluation of the spline calculations.
}
\keyword{internal}
