% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_identifiability.R
\name{check_identifiability}
\alias{check_identifiability}
\title{Test for identifiability of a model}
\usage{
check_identifiability(mod, tol = 0.001)
}
\arguments{
\item{mod}{\code{mod}-object}

\item{tol}{A real; describes the tolerance in the check. If the determinant changes less than 'tol', the system might be not identifiable.
By default  \code{tol = 0.001}}
}
\value{
boolean ok: FALSE if a reason for non-identifiability is found.
}
\description{
Function that tests whether a given model specification might be identifiable.
}
